graph-builder/requirements.txt
biopython==1.79
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
verboselogs==1.7
humanfriendly==10.0
idna==3.3
joblib==1.1.0
lxml==4.9.1
networkx==2.8.6
numpy==1.23.2
obonet==0.3.0
pandas==1.4.4
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
wget==3.2
neo4j==5.3.0
